requirements.txt
requests==2.24.0
emoji==1.2.0
protobuf==3.17.1
PyYAML==5.4.1
Flask==0.10.1
google-cloud-storage
